backend/requirements.txt
flask
flask-cors
python-dotenv
requests
sentence-transformers
transformers[torch]
groq